requirements.txt
django
requests
python-dateutil